# api/uspace/uspacedocs_swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for submitting/monitoring jobs to/from an execution machine",
        "title": "Uspace API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8079",
    "basePath": "/api/v1",
    "paths": {
        "/admin/app": {
            "get": {
                "description": "Admin operations for applications: retrieve, insert, update, delete.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "apps"
                ],
                "summary": "Application admin endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pagination limit for listing applications",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination offset for listing applications",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated application IDs to fetch or delete",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application name (used with version)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application version (used with name)",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Single application ID to delete",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "description": "Single application object (POST/PUT)",
                        "name": "app",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                        }
                    },
                    {
                        "description": "Multiple application objects (POST)",
                        "name": "apps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with application(s) content or status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., malformed input)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Admin operations for applications: retrieve, insert, update, delete.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "apps"
                ],
                "summary": "Application admin endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pagination limit for listing applications",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination offset for listing applications",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated application IDs to fetch or delete",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application name (used with version)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application version (used with name)",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Single application ID to delete",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "description": "Single application object (POST/PUT)",
                        "name": "app",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                        }
                    },
                    {
                        "description": "Multiple application objects (POST)",
                        "name": "apps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with application(s) content or status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., malformed input)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Admin operations for applications: retrieve, insert, update, delete.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "apps"
                ],
                "summary": "Application admin endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pagination limit for listing applications",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination offset for listing applications",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated application IDs to fetch or delete",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application name (used with version)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application version (used with name)",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Single application ID to delete",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "description": "Single application object (POST/PUT)",
                        "name": "app",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                        }
                    },
                    {
                        "description": "Multiple application objects (POST)",
                        "name": "apps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with application(s) content or status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., malformed input)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Admin operations for applications: retrieve, insert, update, delete.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin",
                    "apps"
                ],
                "summary": "Application admin endpoint",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pagination limit for listing applications",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pagination offset for listing applications",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Comma-separated application IDs to fetch or delete",
                        "name": "ids",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application name (used with version)",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Application version (used with name)",
                        "name": "version",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Single application ID to delete",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "description": "Single application object (POST/PUT)",
                        "name": "app",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                        }
                    },
                    {
                        "description": "Multiple application objects (POST)",
                        "name": "apps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/kyri56xcaesar_kuspace_internal_utils.Application"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success with application(s) content or status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request (e.g., malformed input)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Method not allowed",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "kyri56xcaesar_kuspace_internal_utils.Application": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "author_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "inserted_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
...
    }
}